[txn-emitter] Refactor txn-emitter to make it more modular
